Emphasize that there is a list of AZs in an Auto Scaling group
Clarified one of the criteria of attaching an EC2 instance to an Auto Scaling group. 

The original statement that says: "The instance is in the same Availability Zone as the Auto Scaling group", seems to convey that an Auto Scaling group can ONLY handle instances in the same, or in a single, Availability Zone which is entirely false. Take note that an Auto Scaling group can control instances in one or more Availability Zones as per this document: https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-add-availability-zone.html

That statement might give a false impression that the scope of an Auto Scaling group is limited within a single Availability Zone. I would like to propose to change this to the one below, to emphasize that there is a list of AZs in an Auto Scaling group:

"The Availability Zone of the instance is in the list of Availability Zones declared in the Auto Scaling group."