Update attach-instance-asg.md
Attaching a EC2 instance to a Auto Scaling group is a manual process, so I don’t think it is correct to describe it as “an option to enable automatic scaling” 
In the second sentence it is better to use the singular, and the word “considered” is a little better than “become”. 
On line #5 the phrase "The instance to attach" is a little awkward. It is slightly better as "For an instance to be attached," 

Thank you for taking the time to review my proposed changes. 
-Brian